dataset/pyproject.toml
[tool.poetry]
name = "dataset"
version = "0.1.0"
description = ""
authors = ["hanoi0126 <ipocchi129@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.11"
diffusers = "^0.30.3"
torch = "^2.4.1"
hydra-core = "^1.3.2"
transformers = "^4.45.1"
accelerate = "^0.34.2"
xformers = "^0.0.28.post1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
